feat: :sparkles: Add React Router and make Home view
