+++ package.json
@@ -33,7 +33,8 @@
 	},
 	"dependencies": {
 		"react": "^18.3.1",
-		"react-dom": "^18.3.1"
+		"react-dom": "^18.3.1",
+		"react-router-dom": "^6.26.2"
 	},
 	"devDependencies": {
 		"@axe-core/playwright": "^4.10.0",
