+++ src/App.tsx
@@ -1,35 +1,9 @@
 import './App.css';
 
-import { useRef, useState } from 'react';
+import Routes from './routes';
 
-import Events from './components/Events';
-import Navbar from './components/Navbar';
-import { normalizeText } from './utils/normalizeText';
-
-// import { useState } from 'react';
-// import { normalizeText } from './utils/normalizeText';
 function App(): JSX.Element {
-	const [searchTerm, setSearchTerm] = useState('');
-	const CONTAINER_REF = useRef(null);
-
-	function handleNavbarSearch(searchedTerm: string): void {
-		// eslint-disable-next-line no-console
-		console.log(CONTAINER_REF.current);
-
-		setSearchTerm(normalizeText(searchedTerm));
-	}
-
-	return (
-		<>
-			<Navbar
-				ref={CONTAINER_REF}
-				isEnabled
-				onSearch={handleNavbarSearch}
-			/>
-			<Events searchTerm={searchTerm} />
-			{/* <SignUpForm dummy="" /> */}
-		</>
-	);
+	return <Routes />;
 }
 
 export default App;
