+++ src/views/Detail/index.tsx
@@ -1,0 +1,3 @@
+const Details = (): JSX.Element => <div>Detail</div>;
+
+export default Details;
