+++ tsconfig.json
@@ -12,10 +12,13 @@
 			"@context/*": ["src/context/*"],
 			"@data/*": ["src/data/*"],
 			"@hooks/*": ["src/hooks/*"],
+			"@routes/*": ["src/routes/*"],
+			"@state/*": ["src/state/*"],
 			"@services/*": ["src/services/*"],
 			"@templates/*": ["src/templates/*"],
 			"@styles/*": ["src/styles/*"],
 			"@utils/*": ["src/utils/*"],
+			"@views/*": ["src/views/*"],
 			"@images/*": ["src/assets/images/*"]
 		}
 	},
